Do not assert on the order of results
This commit corrects for a race condition present in an integration test
which when executed in CI, may return the classes in an unexpected order.

We now allow the classes to be listed in any order, since this is in the
abstract, an unordered list. Maybe we should sort this in the future?